pyproject.toml
[tool.poetry]
name = "indy-dev-tools"
version = "0.1.0"
description = ""
authors = ["qwukker <minor7addfloortom@gmail.com>"]
readme = "README.md"
packages = [{include = "indy_dev_tools"}]
exclude = ["scripts/*"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.0"
python-dotenv = "^1.0.1"
openai = "^1.11.0"
typer = "^0.9.0"
platformdirs = "^4.2.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"
faster-whisper = "^0.10.0"
pillow = "^10.2.0"

[tool.poetry.scripts]
idt = "indy_dev_tools.main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
